context/requirements.txt
anthropic>=0.40.0
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0
yt-dlp>=2024.12.13
playwright>=1.50.0
google-genai>=1.0.0
